fix r67097 with cast
git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@67099 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ file.c
@@ -1126,7 +1126,7 @@ static inline int
 statx(int dirfd, const char *pathname, int flags,
       unsigned int mask, struct statx *statxbuf)
 {
-    return syscall(__NR_statx, dirfd, pathname, flags, mask, statxbuf);
+    return (int)syscall(__NR_statx, dirfd, pathname, flags, mask, statxbuf);
 }
 #   endif
 # endif
